image-segment-eval/requirements.txt
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0
torch>=2.0.0
pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.7.0
open-clip-torch>=2.20.0
tqdm>=4.65.0
lpips>=0.1.4
scikit-image>=0.20.0
scipy>=1.10.0

